refactor tasks
