+++ corehq/apps/hqadmin/tasks.py
@@ -4,16 +4,29 @@ from celery.schedules import crontab
 from celery.task.base import periodic_task
 
 from corehq.apps.hqadmin.models import HistoricalPillowCheckpoint
-from .utils import check_pillows_for_rewind
+from dimagi.utils.logging import notify_exception
+from pillowtop.utils import get_couch_pillow_instances
+from .utils import check_for_rewind
 
 
 @periodic_task(run_every=crontab(hour=0, minute=0), queue='background_queue')
-def pillow_seq_store_task():
-    check_pillows_for_rewind()
-
+def check_pillows_for_rewind():
+    for pillow in get_couch_pillow_instances():
+        checkpoint = pillow.checkpoint
+        has_rewound, historical_seq = check_for_rewind(checkpoint)
+        if has_rewound:
+            notify_exception(
+                None,
+                message='Found seq number lower than previous for {}. '
+                        'This could mean we are in a rewind state'.format(checkpoint.checkpoint_id),
+                details={
+                    'pillow checkpoint seq': checkpoint.get_current_sequence_id(),
+                    'stored seq': historical_seq
+                }
+            )
 
 @periodic_task(run_every=crontab(hour=0, minute=0), queue='background_queue')
-def create_es_snapshot_checkpoints():
+def create_historical_checkpoints():
     today = date.today()
     thirty_days_ago = today - timedelta(days=30)
     HistoricalPillowCheckpoint.create_pillow_checkpoint_snapshots()
