+++ corehq/apps/hqadmin/utils.py
@@ -3,29 +3,12 @@ from itertools import groupby
 
 from django.conf import settings
 from django.utils.safestring import mark_safe
-
-from corehq.apps.hqadmin.models import HistoricalPillowCheckpoint
-from dimagi.utils.logging import notify_exception
-from pillowtop.utils import force_seq_int, get_couch_pillow_instances
 from restkit import Resource
 
+from corehq.apps.hqadmin.models import HistoricalPillowCheckpoint
+from pillowtop.utils import force_seq_int
+
 EPSILON = 10000000
-
-
-def check_pillows_for_rewind():
-    for pillow in get_couch_pillow_instances():
-        checkpoint = pillow.checkpoint
-        has_rewound, historical_seq = check_for_rewind(checkpoint)
-        if has_rewound:
-            notify_exception(
-                None,
-                message='Found seq number lower than previous for {}. '
-                        'This could mean we are in a rewind state'.format(checkpoint.checkpoint_id),
-                details={
-                    'pillow checkpoint seq': checkpoint.get_current_sequence_id(),
-                    'stored seq': historical_seq
-                }
-            )
 
 
 def check_for_rewind(checkpoint):
